backend/requirements.txt
# Core dependencies
fastapi==0.123.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
python-dotenv==1.2.1

# AI and utilities
anthropic==0.75.0
yt-dlp==2025.11.12

# Audio processing - install in specific order
numpy==1.26.4
scipy==1.16.3
llvmlite==0.45.1
numba==0.62.1
soundfile==0.13.1
librosa==0.11.0

# Essentia - using stable version (2.1b6 has metadata issues)
essentia==2.1b5
